bin/requirements.txt
# Dragons-RansomwareMonitoring Dependencies
# ==========================================

# Core
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0

# Web Scraping & Parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0
tldextract>=5.1.0
aiohttp>=3.9.0
aiohttp-socks>=0.8.0
aiofiles>=23.2.0

# Data Processing
pandas>=2.1.0
stix2>=3.0.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# AI Integration (Optional)
openai>=1.3.0

# Notifications & APIs
telethon>=1.33.0
hudsonrockapi>=0.1.0

# Utilities
pycountry>=23.12.0
validators>=0.22.0
psutil>=5.9.0
urllib3>=2.1.0
